% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weed.R
\docType{data}
\name{pkggrpm}
\alias{pkggrpm}
\title{Marijuana Gross Sales (2014-2017)}
\format{
A data frame with 90 observations and 4 variables.
\itemize{
\item{\code{YEAR}}{ Years of marijuana sales (2014-2017))}
\item{\code{MONTH}}{ Months of marijuana sales }
\item{\code{GROSS_SALES_TYPE}}{ Medical or recreational sales of marijuana}
\item{\code{GROSS_SALES}}{ Total gross sales of marijuana (US dollars)}
}
}
\source{
Data collected by City and County of Denver Office of Marijuana Policy and 
provided by City of Denver Open Data Catalog and used in the "MATH 241" 
create an r data package project.  
See also 
< http://data.denvergov.org>
}
\usage{
data(pkggrpm)
}
\description{
A dataset on monthly medical and recreational marijuana sales from 2014 to
 2017 in Denver, Colorado
}
